+++ dsr_bridge/src/dsr_bridge.cpp
@@ -19,11 +19,9 @@
 #include <iterator>
 #include <string>
 
-// ROS
-#include "nav2_util/string_utils.hpp"
-
 // DSR
 #include "dsr_util/qt_executor.hpp"
+#include "dsr_util/helpers.hpp"
 #include "dsr_bridge/dsr_bridge.hpp"
 
 namespace dsr_bridge
@@ -153,7 +151,7 @@ void DSRBridge::node_from_ros_callback(const dsr_msgs::msg::Node::SharedPtr msg)
     return;
   }
 
-  // Create or update the current node
+  // Create update or delete the current node
   if (!msg->deleted) {
     if (auto node = G_->get_node(msg->name); node.has_value()) {
       modify_attributes(node.value(), msg->attributes);
@@ -168,9 +166,8 @@ void DSRBridge::node_from_ros_callback(const dsr_msgs::msg::Node::SharedPtr msg)
           "The node [%s] couldn't be updated in the DSR", msg->name.c_str());
       }
     } else {
-      auto new_node = create_dsr_node(msg->name, msg->type);
-      modify_attributes(new_node.value(), msg->attributes);
-      if (auto id = G_->insert_node(new_node.value()); id.has_value()) {
+      auto new_node = from_msg(*msg);
+      if (auto id = G_->insert_node(new_node); id.has_value()) {
         RCLCPP_DEBUG(
           this->get_logger(),
           "Inserted [%s] node successfully of type [%s] in the DSR",
@@ -181,46 +178,25 @@ void DSRBridge::node_from_ros_callback(const dsr_msgs::msg::Node::SharedPtr msg)
           "The node [%s] couldn't be inserted in the DSR", msg->name.c_str());
       }
     }
-  } else if (msg->deleted) {
-    // Delete the node
-    if (auto node = G_->get_node(msg->name); node.has_value()) {
-      if (G_->delete_node(node.value().name())) {
-        RCLCPP_DEBUG(
-          this->get_logger(),
-          "Deleted [%s] node successfully of type [%s] in the DSR",
-          msg->name.c_str(), msg->type.c_str());
-      } else {
-        RCLCPP_ERROR(
-          this->get_logger(),
-          "The node [%s] couldn't be deleted in the DSR", msg->name.c_str());
-      }
-
-    } else {
-      RCLCPP_ERROR(
-        this->get_logger(),
-        "The node [%s] could not be deleted in the DSR", msg->name.c_str());
-    }
+  } else {
+    delete_node(msg->name);
   }
 }
 
 // DSR callbacks
-void DSRBridge::node_created(std::uint64_t id, const std::string & /*type*/)
+void DSRBridge::node_created(std::uint64_t id, const std::string & type)
 {
   // Filter the edges that comes from the same source
   if (auto dsr_node = G_->get_node(id); dsr_node.has_value()) {
     if (auto source = G_->get_attrib_by_name<source_att>(dsr_node.value());
       (source.has_value() && source == source_))
     {
-      // Create the message
-      auto node_msg = create_msg_node(dsr_node.value().name(), dsr_node.value().type());
-      // Get all the attributes
-      node_msg.attributes = attributes_to_string(dsr_node.value().attrs());
       // Publish the message
-      node_to_ros_pub_->publish(node_msg);
+      node_to_ros_pub_->publish(to_msg(dsr_node.value()));
       RCLCPP_DEBUG(
         this->get_logger(),
-        "The new node [%s] of type [%s] has been published through ROS",
-        node_msg.name.c_str(), node_msg.type.c_str());
+        "The node [%s] of type [%s] has been published through ROS",
+        dsr_node.value().name().c_str(), type.c_str());
     }
   }
   // Retry to insert lost edges
@@ -256,25 +232,22 @@ void DSRBridge::node_created(std::uint64_t id, const std::string & /*type*/)
   }
 }
 
-void DSRBridge::node_attr_updated(uint64_t id, const std::vector<std::string> & att_names)
+void DSRBridge::node_attr_updated(uint64_t id, const std::vector<std::string> & /*att_names*/)
 {
+  // TODO(ajtudela): Replace this with new attr topic
   // Filter the edges that comes from the same source
   if (auto dsr_node = G_->get_node(id); dsr_node.has_value()) {
     if (auto source = G_->get_attrib_by_name<source_att>(dsr_node.value());
       (source.has_value() && source.value() == source_))
     {
-      // Create the message
-      auto node_msg = create_msg_node(dsr_node.value().name(), dsr_node.value().type());
-      // Mark the node as updated
-      node_msg.updated = true;
       // Get all the updated attributes
-      node_msg.attributes = attributes_updated_to_string(dsr_node.value(), att_names);
+      // node_msg.attributes = attributes_updated_to_string(dsr_node.value(), att_names);
       // Publish the message
-      node_to_ros_pub_->publish(node_msg);
+      node_to_ros_pub_->publish(to_msg(dsr_node.value()));
       RCLCPP_DEBUG(
         this->get_logger(),
-        "The updated node [%s] of type [%s] has been published through ROS",
-        node_msg.name.c_str(), node_msg.type.c_str());
+        "The node [%s] of type [%s] has been published through ROS",
+        dsr_node.value().name().c_str(), dsr_node.value().type().c_str());
     }
   }
 }
@@ -289,7 +262,7 @@ void DSRBridge::edge_updated(std::uint64_t from, std::uint64_t to, const std::st
       // Create the message
       auto edge_msg = create_msg_edge(from, to, type);
       // Get all the attributes
-      edge_msg.attributes = attributes_to_string(dsr_edge.value().attrs());
+      edge_msg.attributes = dsr_util::helpers::attributes_to_string(dsr_edge.value().attrs());
       // Publish the message
       edge_to_ros_pub_->publish(edge_msg);
       RCLCPP_DEBUG(
@@ -327,16 +300,12 @@ void DSRBridge::edge_attr_updated(
 
 void DSRBridge::node_deleted_by_node(const DSR::Node & node)
 {
-  // Create the message
-  auto node_msg = create_msg_node(node.name(), node.type());
-  // Mark the node as deleted
-  node_msg.deleted = true;
   // Publish the message
-  node_to_ros_pub_->publish(node_msg);
+  node_to_ros_pub_->publish(to_msg(node, true));
   RCLCPP_DEBUG(
     this->get_logger(),
     "The deleted node [%s] of type [%s] has been published through ROS",
-    node_msg.name.c_str(), node_msg.type.c_str());
+    node.name().c_str(), node.type().c_str());
 }
 
 void DSRBridge::edge_deleted(std::uint64_t from, std::uint64_t to, const std::string & edge_tag)
@@ -353,16 +322,30 @@ void DSRBridge::edge_deleted(std::uint64_t from, std::uint64_t to, const std::st
     edge_msg.parent.c_str(), edge_msg.child.c_str(), edge_msg.type.c_str());
 }
 
-// Converter functions
-std::optional<DSR::Node> DSRBridge::create_dsr_node(std::string name, std::string type)
+DSR::Node DSRBridge::from_msg(const dsr_msgs::msg::Node & msg)
 {
-  if (!node_types::check_type(type)) {
-    throw std::runtime_error("Error, [" + type + "] is not a valid node type");
+  if (!node_types::check_type(msg.type)) {
+    throw std::runtime_error("Error, [" + msg.type + "] is not a valid node type");
   }
+
   DSR::Node new_node;
-  new_node.name(name);
-  new_node.type(type);
+  new_node.name(msg.name);
+  new_node.type(msg.type);
+  modify_attributes(new_node, msg.attributes);
   return new_node;
+}
+
+dsr_msgs::msg::Node DSRBridge::to_msg(const DSR::Node & node, bool deleted)
+{
+  dsr_msgs::msg::Node node_msg;
+  node_msg.header.stamp = this->now();
+  node_msg.header.frame_id = source_;
+  node_msg.name = node.name();
+  node_msg.type = node.type();
+  node_msg.attributes = dsr_util::helpers::attributes_to_string(node.attrs());
+  node_msg.deleted = deleted;
+  node_msg.updated = false;
+  return node_msg;
 }
 
 std::optional<DSR::Edge> DSRBridge::create_dsr_edge(
@@ -385,18 +368,6 @@ std::optional<DSR::Edge> DSRBridge::create_dsr_edge(
   return {};
 }
 
-dsr_msgs::msg::Node DSRBridge::create_msg_node(std::string name, std::string type)
-{
-  dsr_msgs::msg::Node node_msg;
-  node_msg.header.stamp = this->now();
-  node_msg.header.frame_id = source_;
-  node_msg.name = name;
-  node_msg.type = type;
-  node_msg.updated = false;
-  node_msg.deleted = false;
-  return node_msg;
-}
-
 dsr_msgs::msg::Edge DSRBridge::create_msg_edge(
   std::uint64_t from, std::uint64_t to, const std::string & type)
 {
@@ -417,7 +388,7 @@ dsr_msgs::msg::Edge DSRBridge::create_msg_edge(
 
 // Helper functions
 template<typename TYPE>
-void DSRBridge::modify_attributes(TYPE & elem, std::vector<std::string> & att_str)
+void DSRBridge::modify_attributes(TYPE & elem, const std::vector<std::string> & att_str)
 {
   for (unsigned int i = 0; i < att_str.size(); i += 3) {
     std::string att_name = att_str[i];
@@ -425,75 +396,13 @@ void DSRBridge::modify_attributes(TYPE & elem, std::vector<std::string> & att_st
     std::string att_type = att_str[i + 2];
 
     // Add the attribute to the element
-    DSR::Attribute new_att = string_to_attribute(att_value, std::stoi(att_type));
+    DSR::Attribute new_att = dsr_util::helpers::string_to_attribute(att_value, std::stoi(att_type));
     elem.attrs().insert_or_assign(att_name, new_att);
     RCLCPP_DEBUG(
       this->get_logger(),
       "Updating attribute [%s] = [%s] with type [%ld]",
       att_name.c_str(), att_value.c_str(), new_att.value().index());
   }
-}
-
-std::string DSRBridge::attribute_to_string(const DSR::Attribute & att)
-{
-  std::locale::global(std::locale("C"));
-  switch (att.value().index()) {
-    case 0: {
-        return std::get<std::string>(att.value());
-      }
-    case 1: {
-        return std::to_string(std::get<int32_t>(att.value()));
-      }
-    case 2: {
-        return std::to_string(std::get<float>(att.value()));
-      }
-    case 3: {
-        std::string att_str;
-        for (const auto & value : std::get<std::vector<float>>(att.value())) {
-          att_str += std::to_string(value) + std::string(",");
-        }
-        att_str.pop_back();
-        return att_str;
-      }
-    case 4: {
-        return std::get<bool>(att.value()) ? "true" : "false";
-      }
-    case 5: {
-        std::string att_str;
-        for (const auto & value : std::get<std::vector<uint8_t>>(att.value())) {
-          att_str += std::to_string(value) + std::string(",");
-        }
-        att_str.pop_back();
-        return att_str;
-      }
-    case 6: {
-        return std::to_string(std::get<uint32_t>(att.value()));
-      }
-    case 7: {
-        return std::to_string(std::get<uint64_t>(att.value()));
-      }
-    case 8: {
-        return std::to_string(std::get<double>(att.value()));
-      }
-    default:
-      return "";
-  }
-}
-
-std::vector<std::string> DSRBridge::attributes_to_string(
-  const std::map<std::string, DSR::Attribute> & atts)
-{
-  std::vector<std::string> att_vector_str;
-  for (const auto & [att_name, att_value] : atts) {
-    std::string att_str = attribute_to_string(att_value);
-    att_vector_str.push_back(att_name);
-    att_vector_str.push_back(att_str);
-    att_vector_str.push_back(get_type_from_attribute(att_value));
-    RCLCPP_DEBUG(
-      this->get_logger(),
-      "Attribute [%s] = [%s]", att_name.c_str(), att_str.c_str());
-  }
-  return att_vector_str;
 }
 
 template<typename TYPE>
@@ -505,72 +414,16 @@ std::vector<std::string> DSRBridge::attributes_updated_to_string(
   for (const auto & att_name : atts) {
     auto search = elem.attrs().find(att_name);
     if (search != elem.attrs().end()) {
-      std::string att_value = attribute_to_string(search->second);
+      std::string att_value = dsr_util::helpers::attribute_to_string(search->second);
       att_vector_str.push_back(att_name);
       att_vector_str.push_back(att_value);
-      att_vector_str.push_back(get_type_from_attribute(search->second));
+      att_vector_str.push_back(dsr_util::helpers::get_type_from_attribute(search->second));
       RCLCPP_DEBUG(
         this->get_logger(),
         "Attribute updated [%s] = [%s]", att_name.c_str(), att_value.c_str());
     }
   }
   return att_vector_str;
-}
-
-DSR::Attribute DSRBridge::string_to_attribute(const std::string & att_value, int att_type)
-{
-  DSR::Attribute new_att;
-  switch (att_type) {
-    case 0: {
-        new_att.value(std::string(att_value));
-        break;
-      }
-    case 1: {
-        new_att.value(std::stoi(att_value));
-        break;
-      }
-    case 2: {
-        if (att_value == "nan") {
-          new_att.value(std::numeric_limits<float>::quiet_NaN());
-        } else {
-          new_att.value(std::stof(att_value));
-        }
-        break;
-      }
-    case 3: {
-        std::vector<std::string> values = nav2_util::split(att_value, ',');
-        std::vector<float> float_values;
-        for (const auto & value : values) {
-          if (value == "nan") {
-            float_values.push_back(std::numeric_limits<float>::quiet_NaN());
-          } else {
-            float_values.push_back(std::stof(value));
-          }
-        }
-        new_att.value(float_values);
-        break;
-      }
-    case 4: {
-        new_att.value(att_value == "true");
-        break;
-      }
-    case 6: {
-        new_att.value(std::stoull(att_value));
-        break;
-      }
-    case 7: {
-        new_att.value(std::stoull(att_value));
-        break;
-      }
-    default:
-      break;
-  }
-  return new_att;
-}
-
-std::string DSRBridge::get_type_from_attribute(const DSR::Attribute & att)
-{
-  return std::to_string(att.value().index());
 }
 
 }  // namespace dsr_bridge